Splitted networks and packageconfig for TAO choice.
When working in a network that does not have an Internet connection,
it's not possible to let the configure script to download the TAO archive.
And air gapped networks are not that uncommon at places where OpenDDS
is used.

Added the choice in Yocto to configure OpenDDS to work with the
TAO of doc_group.  Default the OCI version is used.
When added 'PACKAGECONFIG += "doc-group' to the opendds_%.bbappend,
the DOC group version is used.

To be able to use this in an air gapped network, the opendds_git.bb
files has to be masked for bitbake.
This can be done by adding in the local.conf file
'BBMASK += "meta-opendds/recipes-conectivity/opendds/opendds_git.bb'.
In git recipe, there is some nifty parsing to find the latest version.
But this is giving a parsing error when having not external network.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>