opendds: version bump 3.27.0 -> 3.28.0: Update the README file
The latest Long Term Support release of Yocto is Scathgap (5.0).
Supported until April 2028.
The Kirkstone release (4.0) is still supported until April 2026.
Both LTS releases are supported in this meta layer.
But only the latest LTS in the README file should be mentioned.

And update the current version number of OpenDDS in the README.md file

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>