ptest fixes to run the tests on a target system.
All tests are now passing on a Raspberry Pi 4 (32 & 64bit).
Except tests/DCPS/MetaStruct/MetaStructTest what is reported as	#4621.

The tests that should not run on the target are ignored with !CROSS_COMPILE.
These are added to the list of tests with the 0010-WIP-test-filter-tests-that-should-not-run-on-the-tar.patch.
There is no reason (yet) up upstream this.

The Yocto image with the tests can be built with:
  kas build kas/rpi.yaml:kas/ptest.yaml:kas/local.yaml

To run only the opendds tests use:
  ptest-runner opendds

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>